refactor: move some content to a common root module
