+++ honeycomb-core/src/cells/mod.rs
@@ -6,13 +6,3 @@
 
 pub mod collections;
 pub mod orbits;
-
-// ------ CONTENT
-
-/// Type definition for dart identifiers
-///
-/// This is used for better control over memory usage and ID encoding.
-pub type DartIdentifier = u32;
-
-/// Null value for dart identifiers
-pub const NULL_DART_ID: DartIdentifier = 0;
