+++ honeycomb-core/src/lib.rs
@@ -31,8 +31,8 @@
 mod attributes;
 mod cells;
 mod cmap2;
+mod common;
 mod spatial_repr;
-
 #[cfg(feature = "utils")]
 pub mod utils;
 
@@ -49,12 +49,11 @@ pub use cells::{
         NULL_VOLUME_ID,
     },
     orbits::{Orbit2, OrbitPolicy},
-    DartIdentifier, NULL_DART_ID,
 };
-pub use cmap2::{CMap2, CMapError};
+pub use cmap2::CMap2;
+pub use common::{CMapError, CoordsFloat, DartIdentifier, FloatType, NULL_DART_ID};
 pub use spatial_repr::{
     coords::{Coords2, CoordsError},
     vector::Vector2,
     vertex::Vertex2,
-    CoordsFloat, FloatType,
 };
